+++ api/routes/v1/audios/controller/index.js
@@ -1,0 +1,6 @@
+const uploadFile = require('./uploadFile');
+const deleteFile = require('./deleteFile');
+const parseForm = require('./parseForm');
+const preprocess = require('./preprocess');
+
+module.exports = { uploadFile, deleteFile, parseForm, preprocess };
