+++ api/routes/v1/audios/index.js
@@ -1,5 +1,5 @@
 const express = require('express');
-const controller = require('./audios.controller');
+const controller = require('./controller/index');
                        
 const router = express.Router();
